# docs/swagger.yaml
basePath: /api
definitions:
  request.ProductRequestDTO:
    properties:
      name:
        type: string
      price:
        type: number
    required:
    - name
    - price
    type: object
  response.ProductResponseDTO:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Teclado Mecânico
        type: string
      price:
        example: 499.9
        type: number
    type: object
host: localhost:8080
info:
  contact: {}
  description: API de produtos com CRUD genérico e Oracle
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Product API
  version: "1.0"
paths:
  /products:
    get:
      description: Retorna lista de produtos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.ProductResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar produtos
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Cria um novo produto
      parameters:
      - description: Produto a ser criado
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/request.ProductRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ProductResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar produto
      tags:
      - Products
  /products/{id}:
    delete:
      description: Remove um produto pelo ID
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deletar produto
      tags:
      - Products
    get:
      description: Retorna um produto específico pelo ID
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Buscar produto por ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um produto
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: integer
      - description: Dados do produto
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/request.ProductRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualizar produto
      tags:
      - Products
swagger: "2.0"